Add ring-reload-db.  Updates to ring-load.
Add ring-reload-db even though it operates on the old ring policy
of storing pictures in mysql.  This is mostly to give the remediation
script a place to live during the transistion.

Minor updates to ring-load.  Use the simpler readdir which does not
recurse through directories.  Take multiple directories as command
line arguments rather than one at time through a switch.