Improve error handling in ring-resize, improve queue status processing
When there is a failure in ring-resize update the processing queue.  This
may mean that a new queue record is created.  This should ensure that
resize failures never fall completely on the floor.

Detect zero size raw images and set the queue error text.  This is an
improvement over the current segfault.

Remove duplicate queue update from ring-db2files.