Add camera_date
Stop setting date_taken.  This is a historial field and should not
ever be changed.

Stop updating the picture_date automatically.  This date is created
when the picture is first loaded and only manually updated after
that.

Add the camera date. This field either is set from today's date for
from the picture meta data.