Update display.php to include a unique query string
The unique string is required to over come caching of old pictures when
there are edits to the picture without the name of the picture file
changing.  The string is not always unique.  Where possible it is based
on a date_last_maint to allow caches to work in some places.