Add duplicate checking to picture_maint
Move the duplicate checking SQL to the db functions include file
to allow the use of the same logic in both duplicate check and
picture maintenance.

Update the changelog to start a new release cycle.

Remove the old ring-convert-db and ring-db2files since it they are no
longer required and don't match the current database or update
policies.